backend/requirements.txt
# Web Framework
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.12

# Database
sqlalchemy==2.0.35
alembic==1.14.0
psycopg2-binary==2.9.10
asyncpg==0.30.0

# Redis & Caching
redis==5.0.1
hiredis==2.3.2

# Celery (Background Tasks)
celery==5.3.6
flower==2.0.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
cryptography==42.0.0
PyNaCl==1.5.0

# HTTP Client
httpx==0.27.2
aiohttp==3.11.9

# Web Automation
playwright==1.49.1
selenium==4.27.1

# OCR
pytesseract==0.3.10
Pillow==11.0.0
pdf2image==1.17.0

# File Processing
python-magic==0.4.27
PyPDF2==3.0.1

# Validation & Parsing
pydantic[email]==2.10.3
pydantic-settings==2.6.1
email-validator==2.2.0
phonenumbers==8.13.49

# Date & Time
python-dateutil==2.8.2
pytz==2023.3

# Encryption & Hashing
argon2-cffi==23.1.0

# AWS S3 (optional)
boto3>=1.34
botocore>=1.34

# Excel/CSV Processing
openpyxl==3.1.5
pandas==2.2.3
numpy==2.1.3

# XML Processing (for CFDI)
lxml>=5.3.0
xmltodict==0.13.0

# Monitoring & Logging
sentry-sdk==1.39.2
python-json-logger==2.0.7

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-cov==6.0.0
faker==33.1.0

# Frontend (Streamlit)
streamlit==1.40.1

# Code Quality
black==24.10.0
flake8==7.1.1
isort==5.13.2
mypy==1.13.0

# Environment
python-dotenv==1.0.0

# CORS
fastapi-cors==0.0.6

# Rate Limiting
slowapi==0.1.9

# Job Scheduling
apscheduler==3.10.4

# QR Code (for CFDI)
qrcode[pil]==8.0

# Templating
jinja2==3.1.3

# SMS/Email
twilio==8.11.1
sendgrid==6.11.0

# Utils
requests==2.31.0
urllib3>=1.26,<3
